requirements.txt
asyncpg==0.29.0
alembic==1.13.1
fastapi==0.110.1
pytest==8.1.1
redis==5.0.3
SQLAlchemy==2.0.29
httpx==0.27.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
passlib==1.7.4
PyJWT==2.8.0


